+++ lib/crucible_bumblebee/serving.ex
@@ -4,5 +4,10 @@ defmodule CrucibleBumblebee.Serving do
   """
 
   @derive Jason.Encoder
-  defstruct serving_ref: nil, model_ref: nil, surface: nil, metadata: %{}
+  defstruct serving_ref: nil,
+            model_ref: nil,
+            surface: nil,
+            compiled_taps: nil,
+            predict_fun: nil,
+            metadata: %{}
 end